Fixes for lx2160a build
Some OP-TEE related fixes are needed for the platform to boot correctly
using OP-TEE.
Specifically in TF-A for lx2, the option for parsing header on optee is
not supported.  It expects a binary which can be loaded in memory where
execution can start from.

So add an option to support a single binary creation.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>